Don't load config with init function
It was hard to plug this on another package, the main point is that config is loaded with usage of `init` and have `os.Exit(1)` inside this function.
`init` was removed in favor of `LoadCachetConfigClassic` which have to be called to load the config.
A `LoadEmptyConfig` function has been created to initialize a `CachetConfig` without loading data from files (web or local).
Now a project can use this project with the possibility to set the config by itself.